apigateway_cognito/requirements.txt
attrs==19.3.0
aws-cdk.assets==1.15.0
aws-cdk.aws-apigateway==1.15.0
aws-cdk.aws-applicationautoscaling==1.15.0
aws-cdk.aws-autoscaling-common==1.15.0
aws-cdk.aws-certificatemanager==1.15.0
aws-cdk.aws-cloudformation==1.15.0
aws-cdk.aws-cloudwatch==1.15.0
aws-cdk.aws-cognito==1.15.0
aws-cdk.aws-dynamodb==1.15.0
aws-cdk.aws-ec2==1.15.0
aws-cdk.aws-elasticloadbalancingv2==1.15.0
aws-cdk.aws-events==1.15.0
aws-cdk.aws-iam==1.15.0
aws-cdk.aws-kms==1.15.0
aws-cdk.aws-lambda==1.15.0
aws-cdk.aws-logs==1.15.0
aws-cdk.aws-route53==1.15.0
aws-cdk.aws-s3==1.15.0
aws-cdk.aws-s3-assets==1.15.0
aws-cdk.aws-sns==1.15.0
aws-cdk.aws-sqs==1.15.0
aws-cdk.aws-ssm==1.15.0
aws-cdk.core==1.15.0
aws-cdk.cx-api==1.15.0
aws-cdk.region-info==1.15.0
cattrs==0.9.0
jsii==0.19.0
publication==0.0.3
python-dateutil==2.8.1
six==1.13.0
typing-extensions==3.7.4.1
